requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
pypdf
python-dotenv
chromadb
sentence-transformers
fastapi
uvicorn
fpdf
